Implemented a safe routing part1
