+++ backend/db/db_connect.js
@@ -1,0 +1,15 @@
+const mysql = require("mysql");
+
+const connection = mysql.createConnection({
+  host: "localhost",
+  user: "root",
+  password: "",
+  database: "drsrs_attachees",
+});
+
+connection.connect((err) => {
+  if (err) throw err;
+  console.log("Connected to the database!");
+});
+
+module.exports = connection;
