+++ backend/package.json
@@ -10,7 +10,9 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
+    "body-parser": "^1.20.2",
     "express": "^4.19.2",
+    "mysql": "^2.18.1",
     "nodemon": "^3.1.4"
   }
 }
