+++ code/public/index.css
@@ -1,0 +1,191 @@
+
+*{
+    margin: 0;
+    box-sizing: border-box;
+    padding: 0;
+    
+}
+body{
+    display: flex;
+    justify-content: center;
+}
+section{
+    place-items: centre;
+    width: 60%;
+   padding: 10px 30px 6px 14px;
+    display: grid;
+    margin-top: 20px;
+    border: 2px solid #858383;
+}
+.img1{
+    display: none;
+}
+h1{
+    padding-top: 20px;
+    padding-bottom: 40px;
+   text-align: center;
+   width: 80%;
+   color: rgba(0, 0, 255, 0.795);
+   letter-spacing: 1.1px;
+   line-height: 36px;
+   margin-left: 20px;
+}
+.input{
+    margin-top:10px;
+    display: flex;
+    flex-direction: column;
+    position: relative;
+    border: 1px solid grey;
+    padding-top: 20px;
+    padding-bottom: 10px;
+    padding-left: 20px;
+    font-size: 18px;
+    color: rgba(85, 82, 82, 0.945);
+    font-weight: 200;
+    border-radius: 4px;
+    letter-spacing: 1.2px;
+    width: 90%;
+    max-width: 500px;
+    
+}
+.input-box{
+    display: flex;
+    align-items: center;
+    width: 90%;
+    max-width: 500px;
+    padding: 10px 20px;
+    margin-top: 20px;
+    border-radius: 4px;
+    border: 1px solid grey;
+}
+.input-box input{
+    width: 100%;
+    padding: 6px 0;
+    border: 0;
+    outline: 0;
+    font-size: 16px;
+    color: #555;
+}
+.input-box img{
+    width: 8%;
+    margin-right: 8px;
+    float: right;
+}
+.input input{
+    width: 250px;
+    border: 0;
+    outline: 0;
+    font-size: 16px;
+    font-weight: 300;
+}
+.input:hover{
+    border: 1px solid blue;
+    
+}
+
+label{
+    position: absolute;
+    height: 90%;
+    display: flex;
+    align-items: center;
+    transition:800ms ease;
+    padding-bottom: 32px;
+    border: 0;
+    outline: 0;
+    
+}
+input:focus + label,
+input:not(:placeholder-shown)+label{
+    transform: translateY(-18px);
+}
+.log{
+    margin-top: 20px;
+    width: 90%;
+    padding: 8px 4px;
+    border: 0;
+    outline: 0;
+    border-radius: 10px;
+    background-color: rgb(139, 139, 8);
+    color: white;
+    font-size: 20px;
+    cursor: pointer;
+}
+.log:hover{
+       background-color: rgb(167, 167, 6);
+}
+.border{
+    border: 1px solid black;
+    margin-top: 20px;
+}
+a{
+    text-decoration: none;
+    color: white;
+}
+.f{
+    color: blue;
+    cursor: pointer;
+}
+p{
+    margin-left: 2px;
+    margin-top: 30px;
+    color: rgb(0, 0, 0);
+    font-weight: 400;
+    font-size: 20px;
+    width:88%;
+    display: inline-block;
+}
+.sign{
+    background-color: blue;
+    padding: 6px 10px;
+    border: 0;
+    outline: 0;
+    color: white;
+    font-size: 20px;
+    cursor: pointer;
+    float: right;
+    border-radius: 6px;
+}
+.sign:hover{
+       background-color: rgba(0, 0, 255, 0.781);
+}
+
+
+@media(max-width: 700px){
+
+
+}
+
+@media(min-width: 700px){
+
+    section{
+      width: 50%;
+    }
+}
+
+
+
+
+@media(min-width: 1000px){
+    section{
+        width: 29%;
+        margin-left: 120px;
+        padding-top: 10px;
+        padding-bottom: 20px;
+        height: 540px;
+        padding-left: 40px;
+    }
+    .img1{
+    display: flex;
+    width: 30%;
+    margin-left: 40px;
+    margin-top: 20px;
+    height: auto;
+    border-radius: 10px;
+}
+    .log{
+        width: 90%;
+    }
+    .sign{
+        padding: 6px 4px;
+    }
+}
